Lazy loading euphonic (#106)
* Adding euphonic models

* Adding model modifications.

* Adding LL for euphonic.

missing: download of the data, proper testing,
detached app.

* Putting the model and widget in the results panel

* Putting back the results panel in the init.

* Bug fix in the q-section reset.

The problem was that the default value of a tl.List(tl.Float())
is tl.Undefined. Now we explicitely reset the default value
of Q0_vec, h_vec and k_vec

* Adding DownloadYamlHdf5Widget

For now, I added it at the end of the main EuphonicBaseResultsWidget.
It will only allow to download phonopy.yaml and fc.hdf5. The plots can be downloaded in the corresponding tabs.

I added a method to the model which produce the downloadable files, and left the actual
_download_data method in the widget. I think this is a proper design as the model manipulates data,
the view can provide them.

Still missing the download of single tabs, for now it does not work