Refactoring Euphonic MVC. (#108)
* Refactoring Euphonic MVC.

unification of plot and settings widgets, as well as polymorphism instead of inheritance
and composition. This will help maintaining the code clear and readable.

This is still a draft commit, let's say. I started doing the unification in the euphonicmodel.py
and in the structurefactorwidget.py - but the files, imports and everything need to be fixed

* qplanes mod

* more refactoring

* Adding the model and view to the rendering.

Putting init in the model, to inject correct traits.

* Tabs introduction.

Missing several things:

- energy units conversion
- qplanes
- downloads.

* Fixed energy units.

* Energy units, also cm-1

* Fixing single crystal and qplanes

* hiding the third tab INS, for now.

* explicitely "use_c" = False.

* Fixing the qplanes plots, and use_c restored

the c extension is needed to speed up the calculation of the
INS spectra.

* Download of phonopy yaml and hdf5 data.

* Fixing the layout

* fixing detached app and upload of files for INS.

* Adding plotting scripts for INS data.

* Fixed the INS q_planes plots and download

* Update pyproject.toml

otherwise we cannot submit any vibro calc.

* Update pyproject.toml

* Update pyproject.toml

* Setup_phonopy with updated localhost

now local computer is `localhost-hq`